Protocol change to allow any keywork. Finish non-blocking endpoint.
Endpoint requires a listening address and port.  Endpoint is used for
both client and server.  It does not uses the same socket for server
responses, as this was intended to also be able to be used for responses
with long processing time in between and I am under the impression that
a socket shouldn't be held open if unused, rather just create a new one
when the response is needed.

There is no default way to monitor the receiving queue; another
application level thread could watch it, or the handle method could be
extended to send a single, etc.